重构运动参数为双精度浮点类型
将运动参数从整数改为双精度浮点类型（mm/s 和 mm/s²），提高精度和可读性。更新模型、视图模型和 UI 以支持新单位。添加辅助方法将 mm/s 和 mm/s² 转换为脉冲单位，适配硬件接口。优化回零逻辑，删除旧的脉冲相关属性，确保代码一致性和可维护性。

+++ Fredy/Windows/ScanParametersWin/ScanViewModel.cs
@@ -860,7 +860,7 @@ namespace Fredy.Drilling.Holes.ViewModels
              }
              Params.WorkpieceDiameter = diameter;
 
-             Params.SettleTime = Math.Max(1, config.HomeSearchSpeed / 10);
+             Params.SettleTime = Math.Max(1, (int)(config.HomeSearchSpeed / 10));
             Params.DetectMinRadius = config.CircleMinRadius;
             Params.DetectMaxRadius = config.CircleMaxRadius;
             Params.DetectParam1 = config.CircleParam1;
